get user products

+++ src/main/resources/db/migration/V2__Create_products.sql
@@ -5,6 +5,6 @@ create table products.products
     external_id  varchar(255) not null,
     name         varchar(255) not null,
     type         int4         not null,
-    user_id      int4         not null,
+    user_id      bigint       not null,
     primary key (id)
 )
